+++ lib/rofl_ofdpa_fm_driver.cpp
@@ -389,7 +389,7 @@ cofflowmod rofl_ofdpa_fm_driver::disable_tmac_ipv6_unicast_mac(
 }
 
 cofflowmod rofl_ofdpa_fm_driver::enable_ipv4_unicast_host(
-    uint8_t ofp_version, const caddress_in4 &dst, uint32_t group, bool update, uint32_t vrf_id) {
+    uint8_t ofp_version, const caddress_in4 &dst, uint32_t group, bool update, uint16_t vrf_id) {
   cofflowmod fm(ofp_version);
 
   fm.set_command(update ? OFPFC_MODIFY : OFPFC_ADD);
@@ -445,7 +445,7 @@ cofflowmod rofl_ofdpa_fm_driver::enable_ipv4_unicast_host(
 }
 
 cofflowmod rofl_ofdpa_fm_driver::disable_ipv4_unicast_host(
-    uint8_t ofp_version, const caddress_in4 &dst, uint32_t vrf_id) {
+    uint8_t ofp_version, const caddress_in4 &dst, uint16_t vrf_id) {
 
   cofflowmod fm(ofp_version);
 
@@ -475,7 +475,7 @@ cofflowmod rofl_ofdpa_fm_driver::disable_ipv4_unicast_host(
 
 cofflowmod rofl_ofdpa_fm_driver::enable_ipv4_unicast_lpm(
     uint8_t ofp_version, const caddress_in4 &dst, const caddress_in4 &mask,
-    uint32_t group, bool update, uint32_t vrf_id) {
+    uint32_t group, bool update, uint16_t vrf_id) {
   cofflowmod fm(ofp_version);
 
   fm.set_command(update ? OFPFC_MODIFY : OFPFC_ADD);
@@ -531,7 +531,7 @@ cofflowmod rofl_ofdpa_fm_driver::enable_ipv4_unicast_lpm(
 
 cofflowmod rofl_ofdpa_fm_driver::disable_ipv4_unicast_lpm(
     uint8_t ofp_version, const caddress_in4 &dst, const caddress_in4 &mask,
-    uint32_t vrf_id) {
+    uint16_t vrf_id) {
   cofflowmod fm(ofp_version);
 
   fm.set_command(OFPFC_DELETE_STRICT);
@@ -559,7 +559,7 @@ cofflowmod rofl_ofdpa_fm_driver::disable_ipv4_unicast_lpm(
 }
 
 cofflowmod rofl_ofdpa_fm_driver::enable_ipv6_unicast_host(
-    uint8_t ofp_version, const caddress_in6 &dst, uint32_t group, bool update, uint32_t vrf_id) {
+    uint8_t ofp_version, const caddress_in6 &dst, uint32_t group, bool update, uint16_t vrf_id) {
   cofflowmod fm(ofp_version);
 
   fm.set_command(update ? OFPFC_MODIFY : OFPFC_ADD);
@@ -615,7 +615,7 @@ cofflowmod rofl_ofdpa_fm_driver::enable_ipv6_unicast_host(
 }
 
 cofflowmod rofl_ofdpa_fm_driver::disable_ipv6_unicast_host(
-    uint8_t ofp_version, const caddress_in6 &dst, uint32_t vrf_id) {
+    uint8_t ofp_version, const caddress_in6 &dst, uint16_t vrf_id) {
 
   cofflowmod fm(ofp_version);
 
@@ -645,7 +645,7 @@ cofflowmod rofl_ofdpa_fm_driver::disable_ipv6_unicast_host(
 
 cofflowmod rofl_ofdpa_fm_driver::enable_ipv6_unicast_lpm(
     uint8_t ofp_version, const caddress_in6 &dst, const caddress_in6 &mask,
-    uint32_t group, bool update, uint32_t vrf_id) {
+    uint32_t group, bool update, uint16_t vrf_id) {
   cofflowmod fm(ofp_version);
 
   fm.set_command(update ? OFPFC_MODIFY : OFPFC_ADD);
@@ -701,7 +701,7 @@ cofflowmod rofl_ofdpa_fm_driver::enable_ipv6_unicast_lpm(
 
 cofflowmod rofl_ofdpa_fm_driver::disable_ipv6_unicast_lpm(
     uint8_t ofp_version, const caddress_in6 &dst, const caddress_in6 &mask,
-    uint32_t vrf_id) {
+    uint16_t vrf_id) {
   cofflowmod fm(ofp_version);
 
   fm.set_command(OFPFC_DELETE_STRICT);
